+++ simulator/README.md
@@ -1,0 +1,11 @@
+# Simulation with autonomous boats
+
+sim1 :  open loop constant control command
+sim2 :  basic controllers
+sim3 :  pose tracking (lissajou trajectory)
+sim4 :  formation control
+sim5 :  deteriorated formation control
+
+boat.py :  boat class and display class
+controller.py :  controller functions
+saturation.py :  illustrate the saturation of the boat speed
